fixed skybox

+++ shaders/chunk_shader.vert
@@ -9,7 +9,7 @@ layout(set = 0, binding = 0) uniform GlobalUbo {
     mat4 view;
     mat4 projection;
     mat4 projectionView;
-    int dayTime;
+    float dayTime;
 } ubo;
 
 layout(push_constant) uniform Push {
