+++ shaders/skybox.vert
@@ -7,7 +7,7 @@ layout(set = 0, binding = 0) uniform GlobalUbo {
     mat4 view;
     mat4 projection;
     mat4 projectionView;
-    int dayTime;
+    float dayTime;
 } ubo;
 
 layout(location = 0) out vec4 outColor;
